+++ common/src/main/java/com/alien/common/AlienEvents.java
@@ -1,0 +1,10 @@
+package com.alien.common;
+
+import com.alien.common.registry.AlienLifecycleRegistry;
+
+public class AlienEvents {
+
+    public static void onTagsUpdated() {
+        AlienLifecycleRegistry.rebuildLookupMappings();
+    }
+}